move components-related file to components dir
