+++ applications/plugins/SofaCUDA/src/SofaCUDA/component/collision/detection/intersection/CudaProximityIntersection.cpp
@@ -20,11 +20,7 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 #include <sofa/gpu/cuda/CudaTypes.h>
-#include <SofaCUDA/component/solidmechanics/spring/CudaSpringForceField.inl>
-#include <SofaCUDA/component/statecontainer/CudaMechanicalObject.inl>
-#include <SofaCUDA/component/mapping/linear/CudaIdentityMapping.inl>
-#include "CudaContactMapper.h"
-#include <SofaCUDA/component/collision/response/contact/CudaPenalityContactForceField.h>
+#include <SofaCUDA/component/collision/response/mapper/CudaContactMapper.h>
 #include <SofaCUDA/component/collision/geometry/CudaSphereModel.h>
 #include <SofaCUDA/component/collision/geometry/CudaPointModel.h>
 
@@ -33,23 +29,11 @@
 #include <sofa/component/collision/detection/intersection/NewProximityIntersection.inl>
 #include <sofa/component/collision/detection/intersection/MeshNewProximityIntersection.inl>
 #include <sofa/component/collision/detection/intersection/RayDiscreteIntersection.inl>
-#include <sofa/component/collision/detection/intersection/NewProximityIntersection.inl>
-#include <sofa/component/collision/detection/intersection/DiscreteIntersection.h>
 
 #include <sofa/component/collision/response/contact/RayContact.h>
-#include <sofa/component/collision/response/contact/BarycentricPenalityContact.inl>
-#include <sofa/component/collision/response/contact/PenalityContactForceField.h>
-#include <sofa/component/collision/response/mapper/BarycentricContactMapper.inl>
 
-#include <sofa/component/solidmechanics/spring/VectorSpringForceField.h>
 
-#include <sofa/gl/gl.h>
 #include <sofa/helper/Factory.inl>
-#include <sofa/core/Mapping.inl>
-#include <fstream>
-
-
-
 
 namespace sofa::gpu::cuda
 {
