+++ applications/plugins/SofaCUDA/src/SofaCUDA/component/collision/response/mapper/CudaContactMapper.cu
